+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Babashka [http-client](https://github.com/babashka/http-client): HTTP client for Clojure and babashka built on java.net.http
 
+## Unreleased
+
+- [#49](https://github.com/babashka/http-client/issues/49): add `::oauth-token` interceptor
+
 ## 0.4.16 (2024-02-10)
 
 - [#45](https://github.com/babashka/http-client/issues/45): query param values are double encoded